# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Bookstore API",
        "description": "這是書店的 API 文件",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/scrapebooks/GetScrapeProgress": {
            "get": {
                "tags": [
                    "ScrapeBooks"
                ],
                "summary": "獲取爬蟲進度",
                "operationId": "5cc4880db4bbc7241a5bfec626acfe5e",
                "responses": {
                    "200": {
                        "description": "成功返回爬蟲進度",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "progress": {
                                            "properties": {
                                                "current": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 100
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "爬蟲正在進行中"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "獲取進度失敗"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}